refactor: extract 3D visualization backends from visualization.py
Move Plotly 3D, matplotlib 3D, and shared geometry helpers into
separate modules under utils/visualizations/. Reduces visualization.py
from 829 to 260 lines as a slim dispatcher.
